Schema change suggestions
1) For address::get i think it'd be better for the client to have an upvote and downvote count, instead of sending along each vote. That way we can scale without sending tons of notes, and avoid the client having to tally all of the vote counts each time.

Also, do we want to send get address responses sorted in order of (upvote count - downvote count)? I initially thought this sorting should happen in the client, but if the client sorts, then pagination becomes hard because page 2 might need to be sorted ahead of page 1, but the client won't know this until all pages are sent... so perhaps the get response should include sort by field in the DB query? (I haven't looked at BE code so maybe it already works this way lol)

Also, the client would want to persist a user's nametag suggestion after they create it, even if it isn't the highest voted. So the BE should send an "owned" tag. But we'll need to think of a way to do pagination and be sure that the "owned" nametag is always sent in the first page

2) Should the POST to address path that creates a new nametag return all nametags for that address in the response? Presumably, the client has already called that endpoint with GET so has all nametags already, except for the new one... is it better to re-render all addresses on POST? Instead, I think it'd be better to use the nametags already fetched and have client logic set the now POST'd one on screen. Because if we refetch after a post, then the client's new nametag will only have one upvote and be stuck toward the bottom, so the user will need to find it again after the post. So to make that smooth, when a POST is sent then the client should set the new nametag in the box instead of sorting by votes... but this also means that (after refresh) the client should have visibility into which nametags were sent by that user, hence the owned field in the nametag GET response.